+++ src/Database/ClientEntity.php
@@ -24,10 +24,6 @@ final class ClientEntity extends Entity
 
     public ?bool $storeUserToken = null;
 
-    public ?bool $userBecomeAdmin = null;
-
-    public ?bool $keepUserUpdated = null;
-
     public ?array $config = null;
 
     public ?LoginCollection $logins = null;
@@ -39,6 +35,4 @@ final class ClientEntity extends Entity
     public ?UserTokenCollection $userTokens = null;
 
     public ?ClientRuleCollection $rules = null;
-
-    public ?AclRoleCollection $defaultAclRoles = null;
 }