+++ src/Resources/app/administration/src/module/heptacom-admin-open-auth-client/component/rule-container/index.js
@@ -13,6 +13,10 @@ export default {
             required: true,
             type: Object,
         },
+        action: {
+            required: true,
+            type: Array,
+        },
     },
 
     data() {
@@ -27,7 +31,9 @@ export default {
         },
 
         sortedRules() {
-            return this.client.rules.sort((a, b) => a.position - b.position);
+            return this.client.rules
+                .filter(rule => rule.actionName === this.action.name)
+                .sort((a, b) => a.position - b.position);
         },
     },
 
@@ -52,6 +58,8 @@ export default {
         addRule() {
             const rule = this.ruleRepository.create();
             rule.clientId = this.client.id;
+            rule.actionName = this.action.name;
+            rule.actionConfig = {};
             rule.position = this.client.rules.length;
             this.client.rules.add(rule);
         },